+++ attmap/__init__.py
@@ -1,5 +1,6 @@
 """ Package-scope definitions """
 
+from ._att_map_like import AttMapLike
 from .attmap import AttMap
 from .attmap_echo import AttMapEcho
 from .helpers import *
@@ -10,5 +11,6 @@ from ._version import __version__
 AttributeDict = AttMap
 AttributeDictEcho = AttMapEcho
 
-__all__ = ["AttMap", "AttMapEcho", "AttributeDict", "AttributeDictEcho",
+__all__ = ["AttMapLike", "AttMap", "AttMapEcho",
+           "AttributeDict", "AttributeDictEcho",
            "OrdAttMap", "PathExAttMap", "get_data_lines"]