+++ docs/changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.12] - 2019-05-16
+### Added
+- Export base `AttMapLike`.
+### Changed
+- By default, add trailing newline to YAML rendition of an attmap instance; [Issue 48](https://github.com/pepkit/attmap/issues/48)
+- Better API docs
+### Fixed
+- Do not replace double slash in URL with single slash. See [Issue 46](https://github.com/pepkit/attmap/issues/46)
+
 ## [0.11] - 2019-05-16
 ### Added
 - `get_yaml_lines` to get collection of YAML-ready lines from any attmap